requirements.txt
anthropic==0.94.0
beautifulsoup4==4.14.3
dnspython==2.8.0
eth-account==0.13.7
eth_abi==5.2.0
fastapi==0.135.3
httpx==0.28.1
pillow==12.2.0
pydantic==2.12.5
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
python-multipart==0.0.26
python-whois==0.9.6
requests==2.33.1
uvicorn==0.44.0
web3==7.15.0
